+++ src/main/java/com/example/checklist/service/DepartementService.java
@@ -9,6 +9,6 @@ public interface DepartementService {
     public Optional<Departement> DepartementSave (Departement departement );
     public List<Departement> findAll();
     public void DeleteId(Long id);
-    public Departement UpdateDepartement(Departement departement,int id);
+    public Departement UpdateDepartement(Departement departement, Long id);
     public long countDepartement();
 }
